app/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
click==8.1.7
click-spinner==0.1.10
colorama==0.4.6
croniter==1.4.1
databases==0.9.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.99.0
fastapi-cli==0.0.5
fastapi-utilities==0.2.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
mangum==0.19.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
mysqlclient==2.2.4
numpy==2.1.2
openpyxl==3.1.5
pyactiveresource==2.2.2
pydantic==1.10.18
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
rich==13.8.1
shellingham==1.5.4
ShopifyAPI==12.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.27.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
yarl==1.11.1
